+++ src/hooks/CustomFixtures.ts
@@ -49,9 +49,9 @@ import AccessDeniedPage from '../pages/IRAS/AccessDeniedPage';
 import RtsPage from '../pages/Common/RtsPage';
 import ParticipatingOrganisationsPage from '../pages/IRAS/makeChanges/modifications/ParticipatingOrganisationsPage';
 import SelectAreaOfChangePage from '../pages/IRAS/makeChanges/modifications/SelectAreaOfChangePage';
-import SearchModificationsPage from '../pages/IRAS/reviewResearch/approvals/SearchModificationsPage';
-import ApprovalsPage from '../pages/IRAS/reviewResearch/approvals/ApprovalsPage';
+import SearchModificationsPage from '../pages/IRAS/reviewResearch/receiveAmendments/SearchModificationsPage';
 import ModificationsReadyToAssignPage from '../pages/IRAS/reviewResearch/receiveAmendments/ModificationsReadyToAssignPage';
+import ApprovalsPage from '../pages/IRAS/reviewResearch/approvals/ApprovalsPage';
 
 type CustomFixtures = {
   commonItemsPage: CommonItemsPage;
