+++ tests/features/cross-compatible/accessibility/AxeAccessibility.feature
@@ -850,7 +850,7 @@ Feature: Run Axe Accessibilty Test Tool Against App Pages
     When I enter 'Valid_Iras_Id_Nth' into the search field
     And I select advanced filters in the search modifications page using 'Advanced_Filters_Nth'
     And I click the 'Apply_filters' button on the 'Search_Modifications_Page'
-    Then I can see the selected filters 'Advanced_Filters_Nth' are displayed under active filters
+    Then I can see the selected filters 'Advanced_Filters_Nth' are displayed under active filters in the search modifications page
     And the system displays modification records matching the search 'Valid_Iras_Id_Nth' and filter criteria 'Advanced_Filters_Nth'
     When I Scan the page with the Axe Accessibilty Tool
     And I analyse the results from the Axe scan
