+++ tests/features/cross-compatible/stories/systemTests/reviewResearch/receiveAmendments/AdvancedFilterSearchModifications.feature
